Create a new folder called Enums and a new file for the DisplayMode enum
Improve code estructure
